Conflicts:
	scilab/Scilab.sln
	scilab/configure
	scilab/etc/modules.xml.in
	scilab/etc/modules.xml.vc
	scilab/modules/Makefile.in
	scilab/modules/api_scilab/src/cpp/api_int.cpp
	scilab/modules/call_scilab/LibScilab_Import.def
	scilab/modules/call_scilab/src/c/call_scilab.c
	scilab/modules/console/src/jni/GuiManagement.i
	scilab/modules/console/src/jni/GuiManagement_wrap.c
	scilab/modules/core/Makefile.am
	scilab/modules/core/Makefile.in
	scilab/modules/core/src/c/TerminateCore.c
	scilab/modules/core/src/c/callDynamicGateway.c
	scilab/modules/core/src/cpp/hashtable_core.cpp
	scilab/modules/core/src/fortran/core_f.vfproj
	scilab/modules/elementary_functions/sci_gateway/c/gw_elementary_functions.c
	scilab/modules/elementary_functions/sci_gateway/c/sci_zeros.c
	scilab/modules/fileio/Makefile.am
	scilab/modules/fileio/Makefile.in
	scilab/modules/fileio/sci_gateway/c/sci_fscanfMat.c
	scilab/modules/fileio/sci_gateway/c/sci_mputl.c
	scilab/modules/fileio/src/c/basename.h
	scilab/modules/fileio/src/c/mgetl.c
	scilab/modules/functions/Makefile.in
	scilab/modules/hdf5/sci_gateway/cpp/sci_import_from_hdf5.cpp
	scilab/modules/javasci/Javasci.vcproj
	scilab/modules/jvm/Makefile.am
	scilab/modules/jvm/Makefile.in
	scilab/modules/localization/src/c/charEncoding.c
	scilab/modules/matio/src/c/GetCharVariable.c
	scilab/modules/output_stream/src/c/errmsg.c
	scilab/modules/parallel/Makefile.am
	scilab/modules/parallel/Makefile.in
	scilab/modules/parallel/includes/gw_parallel.h
	scilab/modules/parallel/sci_gateway/c/gw_parallel.c
	scilab/modules/parallel/src/c/parallel.vcproj
	scilab/modules/scinotes/Makefile.in
	scilab/modules/string/sci_gateway/c/sci_strstr.c
	scilab/modules/types/Makefile.am
	scilab/modules/types/Makefile.in
	scilab/modules/ui_data/includes/gw_ui_data.h
	scilab/modules/ui_data/sci_gateway/c/gw_ui_data.c
	scilab/modules/ui_data/sci_gateway/cpp/sci_editvar.cpp
	scilab/modules/ui_data/sci_gateway/cpp/sci_updateEditvarValue.cpp
	scilab/modules/xcos/includes/gw_xcos.h
	scilab/modules/xcos/sci_gateway/c/gw_xcos.c

+++ scilab/modules/xcos/src/java/org/scilab/modules/xcos/block/io/ContextUpdate.java
@@ -21,12 +21,16 @@ import java.util.ArrayList;
 import java.util.EnumMap;
 import java.util.List;
 import java.util.Map;
+import java.util.regex.Pattern;
 
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 import org.scilab.modules.action_binding.highlevel.ScilabInterpreterManagement;
 import org.scilab.modules.action_binding.highlevel.ScilabInterpreterManagement.InterpreterException;
-import org.scilab.modules.types.scilabTypes.ScilabDouble;
-import org.scilab.modules.types.scilabTypes.ScilabList;
-import org.scilab.modules.types.scilabTypes.ScilabType;
+import org.scilab.modules.types.ScilabDouble;
+import org.scilab.modules.types.ScilabList;
+import org.scilab.modules.types.ScilabString;
+import org.scilab.modules.types.ScilabType;
 import org.scilab.modules.xcos.block.BasicBlock;
 import org.scilab.modules.xcos.block.SuperBlock;
 import org.scilab.modules.xcos.graph.SuperBlockDiagram;
@@ -51,6 +55,7 @@ import com.mxgraph.util.mxEventObject;
  */
 public abstract class ContextUpdate extends BasicBlock {
 
+	private static final Log LOG_LOCAL = LogFactory.getLog(ContextUpdate.class);
 	private static final long serialVersionUID = 6076826729067963560L;
 
 	/**
@@ -111,11 +116,57 @@ public abstract class ContextUpdate extends BasicBlock {
 	}
 	
 	/**
+	 * Implement a listener to update the {@link ContextUpdate#isContextDependent} flag.
+	 */
+	private static final class ExprsChangeAdapter implements PropertyChangeListener, Serializable {
+		private static final Pattern INTEGER_PATTERN = Pattern.compile("\\d+");
+
+		private static ExprsChangeAdapter instance;
+		
+		/**
+		 * Default constructor
+		 */
+		public ExprsChangeAdapter() { }
+		
+		/**
+		 * @return the shared instance
+		 */
+		public static ExprsChangeAdapter getInstance() {
+			if (instance == null) {
+				instance = new ExprsChangeAdapter();
+			}
+			return instance;
+		}
+		
+		/**
+		 * isContextDependant field
+		 * 
+		 * @param evt the event
+		 * @see java.beans.PropertyChangeListener#propertyChange(java.beans.PropertyChangeEvent)
+		 */
+		@Override
+		public void propertyChange(final PropertyChangeEvent evt) {
+			final ScilabType data = (ScilabType) evt.getNewValue();
+			final ContextUpdate ioBlock = (ContextUpdate) evt.getSource();
+			
+			if (!data.isEmpty()) {
+				final String newIndex = ((ScilabString) data).getData()[0][0];
+				
+				if (!INTEGER_PATTERN.matcher(newIndex).matches()) {
+					ioBlock.isContextDependent = true;
+				} else {
+					ioBlock.isContextDependent = false;
+				}
+			}
+		}
+	}
+	
+	/**
 	 * This enum represent all the subclasses of ContextUpdate .
 	 * 
 	 * It is used to easily loop over a BasicBlock I/O blocks
 	 */
-	public enum IOBlocks {
+	public static enum IOBlocks {
 		/** Map a control port to an event input block */
 		EventInBlock(EventInBlock.class, ControlPort.class),
 		/** Map a command port to an event output block */
@@ -204,8 +255,8 @@ public abstract class ContextUpdate extends BasicBlock {
 			final Object[] children = mxGraphModel.getChildCells(
 					defaultModel, defaultParent, true, false);
 			
-			for (int i = 0; i < children.length; i++) {
-				final mxICell child = (mxICell) children[i];
+			for (Object element : children) {
+				final mxICell child = (mxICell) element;
 
 				/* if compatible add it to the list */
 				for (IOBlocks b : IOBlocks.values()) {
@@ -233,14 +284,18 @@ public abstract class ContextUpdate extends BasicBlock {
 		}
 	}
 	
+	private transient boolean isContextDependent;
+	
 	/**
 	 * Constructor.
 	 */
 	public ContextUpdate() {
 		super();
 		
-		getParametersPCS().addPropertyChangeListener("integerParameters",
+		getParametersPCS().addPropertyChangeListener(INTEGER_PARAMETERS,
 				IndexChangeAdapter.getInstance());
+		getParametersPCS().addPropertyChangeListener(EXPRS,
+				ExprsChangeAdapter.getInstance());
 	}
 
 	/**
@@ -265,9 +320,16 @@ public abstract class ContextUpdate extends BasicBlock {
 			return;
 		}
 
-		final File tempOutput;
-		final File tempInput;
-		final File tempContext;
+		// do not evaluate context is the block is not context dependent.
+		if (!isContextDependent) {
+			return;
+		}
+		
+		LOG_LOCAL.trace("Update the I/O value from the context");
+		
+		File tempOutput;
+		File tempInput;
+		File tempContext;
 		try {
 			tempInput = FileUtils.createTempFile();
 			tempInput.deleteOnExit();
@@ -290,7 +352,7 @@ public abstract class ContextUpdate extends BasicBlock {
 			updateBlockSettings(modifiedBlock);
 
 		} catch (IOException e) {
-			e.printStackTrace();
+			LOG_LOCAL.error(e);
 		}
 	}
 }
